+++ website/src/theme/CodeSnippet/index.js
@@ -1,12 +1,15 @@
-import React, {useEffect, useState} from 'react';
-
-import styles from './styles.module.css';
-import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+import React, {useEffect, useState, useRef} from 'react';
+import clsx from 'clsx';
+import Highlight, {defaultProps} from 'prism-react-renderer';
+import copy from 'copy-text-to-clipboard';
 import useThemeContext from '@theme/hooks/useThemeContext';
-
-import Highlight, { defaultProps } from "prism-react-renderer";
+import Translate, {translate} from '@docusaurus/Translate';
+import Link from '@docusaurus/Link';
 
 import monokai from "@site/src/plugins/prism_themes/monokai";
+import useDocusaurusContext from '@docusaurus/useDocusaurusContext';
+
+import styles from './styles.module.css';
 
 function CodeSnippet(props) {
   const {
@@ -15,6 +18,7 @@ function CodeSnippet(props) {
     },
   } = useDocusaurusContext();
 
+  const [showCopied, setShowCopied] = useState(false);
   const [mounted, setMounted] = useState(false);
   // The Prism theme on SSR is always the default theme but the site theme
   // can be in a different mode. React hydration doesn't update DOM styles
@@ -32,23 +36,70 @@ function CodeSnippet(props) {
   const darkModeTheme = prism.darkTheme || lightModeTheme;
   const prismTheme = isDarkTheme ? darkModeTheme : lightModeTheme;
 
+  const button = useRef(null);
+
   const {
     lang = 'yaml',
-    snippet,
+    config,
+    copyBit,
+    further,
   } = props;
 
+  let copySnippet = copyBit || config;
+  const handleCopyCode = () => {
+    copy(copySnippet);
+    setShowCopied(true);
+
+    setTimeout(() => setShowCopied(false), 2000);
+  };
+
   return (
-    <Highlight {...defaultProps} key={mounted} theme={prismTheme} code={snippet} language={lang}>
+    <Highlight {...defaultProps} key={mounted} theme={prismTheme} code={config} language={lang}>
       {({ className, style, tokens, getLineProps, getTokenProps }) => (
-        <pre className={`${className} ${styles.codeSnippet}`} style={style}>
-          {tokens.map((line, i) => (
-            <div {...getLineProps({ line, key: i })}>
-              {line.map((token, key) => (
-                <span {...getTokenProps({ token, key })} />
+        <div className={styles.codeBlockContainer}>
+          <div className={clsx(styles.codeBlockContent, lang)}>
+            <pre className={`${className}`} style={style}>
+              {tokens.map((line, i) => (
+                <div {...getLineProps({ line, key: i })}>
+                  {line.map((token, key) => (
+                    <span {...getTokenProps({ token, key })} />
+                  ))}
+                </div>
               ))}
-            </div>
-          ))}
-        </pre>
+            </pre>
+
+            <button
+              ref={button}
+              type="button"
+              aria-label={translate({
+                id: 'theme.CodeBlock.copyButtonAriaLabel',
+                message: 'Copy code to clipboard',
+                description: 'The ARIA label for copy code blocks button',
+              })}
+              className={clsx(styles.copyButton, 'clean-btn')}
+              onClick={handleCopyCode}>
+              {showCopied ? (
+                <Translate
+                  id="theme.CodeBlock.copied"
+                  description="The copied button label on code blocks">
+                  Copied
+                </Translate>
+              ) : (
+                <Translate
+                  id="theme.CodeBlock.copy"
+                  description="The copy button label on code blocks">
+                  Copy
+                </Translate>
+              )}
+            </button>
+
+            {further && <Link
+              className={clsx(styles.furtherButton, 'button button--outline button--primary')}
+              to={further}>
+              Read about
+            </Link>}
+          </div>
+        </div>
       )}
     </Highlight>
   );